Rename argparse to argparser
Under python 3.6 / PyCharm Debugger, the import argparse was getting the local argparse and not the library.
AttributeError: 'module' object has no attribute 'ArgumentParser'
Renaming our local argparse.py to argparser.py prevents this collision
